+++ modules/aws-anyscale-iam/iam-policies-data.tf
@@ -97,10 +97,11 @@ data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
       sid    = "RestrictInstanceStart"
       effect = "Allow"
       actions = [
+        "ec2:StartInstances",
         "ec2:RunInstances"
       ]
       resources = [
-        "arn:aws:ec2:*:${local.account_id}:instance/*",
+        "arn:aws:ec2:*:${local.account_id}:instance/*"
       ]
       condition {
         test     = "StringEquals"
@@ -115,6 +116,26 @@ data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
     }
   }
 
+  dynamic "statement" {
+    for_each = local.cloud_id_provided ? [1] : []
+    content {
+      sid    = "AllowRunInstancesForUntaggedResources"
+      effect = "Allow"
+      actions = [
+        "ec2:RunInstances"
+      ]
+      resources = [
+        "arn:aws:ec2:*::image/*",
+        "arn:aws:ec2:*::snapshot/*",
+        "arn:aws:ec2:*:*:subnet/*",
+        "arn:aws:ec2:*:*:network-interface/*",
+        "arn:aws:ec2:*:*:security-group/*",
+        "arn:aws:ec2:*:*:key-pair/*",
+        "arn:aws:ec2:*:*:volume/*"
+      ]
+    }
+  }
+
 
   dynamic "statement" {
     for_each = local.cloud_id_provided ? [1] : []
@@ -123,8 +144,7 @@ data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
       effect = "Allow"
       actions = [
         "ec2:TerminateInstances",
-        "ec2:StopInstances",
-        "ec2:StartInstances"
+        "ec2:StopInstances"
       ]
       resources = [
         "arn:aws:ec2:*:${local.account_id}:instance/*"
@@ -161,6 +181,14 @@ data "aws_iam_policy_document" "iam_anyscale_steadystate_policy" {
         test     = "StringNotEquals"
         variable = "aws:ResourceTag/anyscale-cloud-id"
         values   = [var.anyscale_cloud_id]
+      }
+      condition {
+        test     = "StringNotEquals"
+        variable = "ec2:CreateAction"
+        values = [
+          "RunInstances",
+          "StartInstances"
+        ]
       }
     }
   }
@@ -243,4 +271,271 @@ data "aws_iam_policy_document" "iam_anyscale_s3_bucket_access" {
       ]
     }
   }
+}
+
+data "aws_iam_policy_document" "iam_anyscale_services_v2" {
+  dynamic "statement" {
+    for_each = local.cloud_id_provided ? [] : [1]
+    content {
+      sid = "Cloudformation"
+      actions = [
+        "cloudformation:TagResource",
+        "cloudformation:UntagResource",
+        "cloudformation:CreateStack",
+        "cloudformation:UpdateStack",
+        "cloudformation:DeleteStack",
+        "cloudformation:DescribeStackEvents",
+        "cloudformation:DescribeStackResources",
+        "cloudformation:DescribeStacks"
+      ]
+      resources = [
+        "arn:aws:cloudformation:*:${local.account_id}:stack/*"
+      ]
+    }
+  }
+  dynamic "statement" {
+    for_each = local.cloud_id_provided ? [1] : []
+    content {
+      sid    = "CloudformationRead"
+      effect = "Allow"
+      actions = [
+        "cloudformation:DescribeStackEvents",
+        "cloudformation:DescribeStackResources",
+        "cloudformation:DescribeStacks"
+      ]
+      resources = [
+        "arn:aws:cloudformation:*:${local.account_id}:stack/*"
+      ]
+    }
+  }
+
+  dynamic "statement" {
+    for_each = local.cloud_id_provided ? [1] : []
+    content {
+      sid    = "CloudformationWrite"
+      effect = "Allow"
+      actions = [
+        "cloudformation:TagResource",
+        "cloudformation:UntagResource",
+        "cloudformation:CreateStack",
+        "cloudformation:UpdateStack",
+        "cloudformation:DeleteStack",
+      ]
+      resources = [
+        "arn:aws:cloudformation:*:${local.account_id}:stack/*"
+      ]
+      condition {
+        test     = "StringEquals"
+        variable = "aws:ResourceTag/anyscale-cloud-id"
+        values   = [var.anyscale_cloud_id]
+      }
+    }
+  }
+
+  statement {
+    sid    = "ELBRead"
+    effect = "Allow"
+    actions = [
+      "elasticloadbalancing:DescribeListeners",
+      "elasticloadbalancing:DescribeLoadBalancers",
+      "elasticloadbalancing:DescribeRules",
+      "elasticloadbalancing:DescribeTargetGroups",
+      "elasticloadbalancing:DescribeTargetHealth",
+      "elasticloadbalancing:DescribeListenerCertificates"
+    ]
+    resources = ["*"]
+  }
+
+  statement {
+    sid    = "ELBCertsWrite"
+    effect = "Allow"
+    actions = [
+      "elasticloadbalancing:AddListenerCertificates",
+      "elasticloadbalancing:RemoveListenerCertificates"
+    ]
+    resources = [
+      "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*"
+    ]
+  }
+
+  statement {
+    sid    = "ACMRead"
+    effect = "Allow"
+    actions = [
+      "acm:ListCertificates"
+    ]
+    resources = ["*"]
+  }
+
+  statement {
+    sid    = "ACMWrite"
+    effect = "Allow"
+    actions = [
+      "acm:DeleteCertificate",
+      "acm:RenewCertificate",
+      "acm:RequestCertificate",
+      "acm:AddTagsToCertificate",
+      "acm:DescribeCertificate",
+      "acm:GetCertificate",
+      "acm:ListTagsForCertificate"
+    ]
+    resources = ["arn:aws:acm:*:${local.account_id}:certificate/*"]
+  }
+
+  statement {
+    sid    = "ELBWrite"
+    effect = "Allow"
+    actions = [
+      "elasticloadbalancing:AddTags",
+      "elasticloadbalancing:CreateRule",
+      "elasticloadbalancing:ModifyRule",
+      "elasticloadbalancing:DeleteRule",
+      "elasticloadbalancing:CreateListener",
+      "elasticloadbalancing:ModifyListener",
+      "elasticloadbalancing:DeleteListener",
+      "elasticloadbalancing:CreateLoadBalancer",
+      "elasticloadbalancing:DeleteLoadBalancer",
+      "elasticloadbalancing:CreateTargetGroup",
+      "elasticloadbalancing:ModifyTargetGroup",
+      "elasticloadbalancing:DeleteTargetGroup",
+      "elasticloadbalancing:RegisterTargets",
+      "elasticloadbalancing:DeregisterTargets"
+    ]
+    resources = [
+      "arn:aws:elasticloadbalancing:*:${local.account_id}:loadbalancer/app/*",
+      "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*",
+      "arn:aws:elasticloadbalancing:*:${local.account_id}:listener-rule/app/*",
+      "arn:aws:elasticloadbalancing:*:${local.account_id}:targetgroup*"
+    ]
+    condition {
+      test     = "StringEquals"
+      variable = "aws:CalledViaFirst"
+      values   = ["cloudformation.amazonaws.com"]
+    }
+  }
+  # dynamic "statement" {
+  #   for_each = local.cloud_id_provided ? [] : [1]
+  #   content {
+  #     sid    = "ELBWrite"
+  #     effect = "Allow"
+  #     actions = [
+  #       "elasticloadbalancing:AddTags",
+  #       "elasticloadbalancing:CreateRule",
+  #       "elasticloadbalancing:ModifyRule",
+  #       "elasticloadbalancing:DeleteRule",
+  #       "elasticloadbalancing:CreateListener",
+  #       "elasticloadbalancing:ModifyListener",
+  #       "elasticloadbalancing:DeleteListener",
+  #       "elasticloadbalancing:CreateLoadBalancer",
+  #       "elasticloadbalancing:DeleteLoadBalancer",
+  #       "elasticloadbalancing:CreateTargetGroup",
+  #       "elasticloadbalancing:ModifyTargetGroup",
+  #       "elasticloadbalancing:DeleteTargetGroup",
+  #       "elasticloadbalancing:RegisterTargets",
+  #       "elasticloadbalancing:DeregisterTargets"
+  #     ]
+  #     resources = [
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:loadbalancer/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener-rule/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:targetgroup*"
+  #     ]
+  #     condition {
+  #       test     = "StringEquals"
+  #       variable = "aws:CalledViaFirst"
+  #       values   = ["cloudformation.amazonaws.com"]
+  #     }
+  #   }
+  # }
+
+  # dynamic "statement" {
+  #   for_each = local.cloud_id_provided ? [1] : []
+  #   content {
+  #     sid    = "ELBWrite"
+  #     effect = "Allow"
+  #     actions = [
+  #       "elasticloadbalancing:AddTags",
+  #       "elasticloadbalancing:DeleteTags",
+  #       "elasticloadbalancing:DeleteLoadBalancer",
+  #       "elasticloadbalancing:ModifyRule",
+  #       "elasticloadbalancing:DeleteRule",
+  #       "elasticloadbalancing:ModifyListener",
+  #       "elasticloadbalancing:DeleteListener",
+  #       "elasticloadbalancing:ModifyTargetGroup",
+  #       "elasticloadbalancing:DeleteTargetGroup",
+  #       "elasticloadbalancing:RegisterTargets",
+  #       "elasticloadbalancing:DeregisterTargets",
+  #       "elasticloadbalancing:CreateTargetGroup"
+  #     ]
+  #     resources = [
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:loadbalancer/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener-rule/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:targetgroup*"
+  #     ]
+  #     condition {
+  #       test     = "StringEquals"
+  #       variable = "aws:CalledViaFirst"
+  #       values   = ["cloudformation.amazonaws.com"]
+  #     }
+  #   }
+  # }
+
+  # dynamic "statement" {
+  #   for_each = local.cloud_id_provided ? [1] : []
+  #   content {
+  #     sid    = "ELBWriteRestricted"
+  #     effect = "Allow"
+  #     actions = [
+  #       "elasticloadbalancing:CreateRule",
+  #       "elasticloadbalancing:CreateListener",
+  #       "elasticloadbalancing:CreateLoadBalancer",
+  #       # "elasticloadbalancing:CreateTargetGroup",
+  #     ]
+  #     resources = [
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:loadbalancer/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener-rule/app/*",
+  #       # "arn:aws:elasticloadbalancing:*:${local.account_id}:targetgroup*"
+  #     ]
+  #     condition {
+  #       test     = "StringEquals"
+  #       variable = "aws:CalledViaFirst"
+  #       values   = ["cloudformation.amazonaws.com"]
+  #     }
+  #     condition {
+  #       test     = "StringEquals"
+  #       variable = "aws:RequestTag/anyscale-cloud-id"
+  #       values   = [var.anyscale_cloud_id]
+  #     }
+  #     condition {
+  #       test     = "ForAnyValue:StringEquals"
+  #       variable = "aws:TagKeys"
+  #       values   = ["anyscale-cloud-id"]
+  #     }
+  #   }
+  # }
+
+  # dynamic "statement" {
+  #   for_each = local.cloud_id_provided ? [1] : []
+  #   content {
+  #     sid    = "DenyTaggingOnOtherLoadbalancers"
+  #     effect = "Deny"
+  #     actions = [
+  #       "elasticloadbalancing:AddTags",
+  #       "elasticloadbalancing:DeleteTags",
+  #     ]
+  #     resources = [
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:loadbalancer/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener/app/*",
+  #       "arn:aws:elasticloadbalancing:*:${local.account_id}:listener-rule/app/*",
+  #       # "arn:aws:elasticloadbalancing:*:${local.account_id}:targetgroup*"
+  #     ]
+  #     condition {
+  #       test     = "StringNotEquals"
+  #       variable = "aws:ResourceTag/anyscale-cloud-id"
+  #       values   = [var.anyscale_cloud_id]
+  #     }
+  #   }
+  # }
 }